fix: remove Volume#server= to attach volumes
new fog-core model#reload resets the model the current remote state.
The previous implementation of volume relied upon initializing with a
device, reloading the model several times, and then attaching with a
server instance is set.  IMO, the new fog-core is doing the correct
thing.

To support this behavior, make #attach and #detach public methods and
require a device argument in #attach.  #server= does exist but only as
an attr_writer.  This may cause some confusion in the future release
and must be included in the CHANGELOG.