Do not require access keys in Storage class
I wish to use an IAM profile for an instance to grant access to S3
buckets rather than pass in access keys.

This is possible throughout fog-aws by passing the `use_iam_profile`,
but the `Storage` class requires aws_access_key_id and
aws_secret_access_key parameters.

This causes errors related to malformed headers being submitted to
Amazon S3 similar to being described in this issue. [1]

As suggested by the issue, workarounds such as passing in blank or bogus
access keys do not resolve the issue.

The error is resolved by avoiding the neccessity to pass in the access
keys and only pass in the use_iam_profile parameter.

[1] https://github.com/fog/fog-aws/issues/423#issuecomment-367133454