Update aurora check to use start_with? instead of ==.
Early versions of Aurora used just "aurora" as the value for the engine parameter. But Aurora now supports more than one database, so the engine string has been updated to, as of this writing, support aurora-mysql and aurora-postgres.

Without this change, it's impossible to use this method to create individual aurora instances because the code fog forces the caller to specify the master_username, password, and allocated_storage values. None of which are valid for Aurora instances.

I also updated the default "flavor_id" values to more recent versions since db.m1s don't exist anymore, and db.r3s are being phased out.