Support FIPS endpoints, disable S3 Transfer Acceleration in GovCloud (#730)
* Support FIPS endpoints, disable S3 Transfer Acceleration in GovCloud

The Fog::AWS::Utils region_to_host method returns the standard S3
endpoints even when ENV['AWS_USE_FIPS_ENDPOINT']=='true'.  When FIPS
is called for, and we are in a region where FIPS endpoints are
available,  this method should return the FIPS endpoint.

Furthermore, when S3 Transfer Acceleration (S3TA) is requested by
configuration, the above endpoint gets overridden to select the S3TA
endpoint.  However, S3TA is not avaialble in GovCloud, and has no FIPS
endpoint equivalents.  In this instance, if the region is a GovCloud
region, or if FIPS mode is called for, do _not_ override the endpoint
to use S3TA.

* lint cleanups

* disable S3TA and warn if in GovCloud or FIPS mode

* Update lib/fog/aws/storage.rb

Fix a typo in acceleration disable warning.

---------

Co-authored-by: Wesley Beary <geemus@gmail.com>